// DocumentKnowledgeMining/App/frontend-app/src/api/apiTypes/documentResults.ts
  export interface Document {
    documentId: string;  // Unique document identifier
    fileName: string;    // Name of the file
    keywords: {          // Keywords object with dynamic keys and comma-separated string values
      [key: string]: string;
    };
    importedTime: string;      // ISO timestamp for when the document was imported
    processingTime: string;    // Time taken to process the document
    mimeType: string;          // MIME type of the document (e.g., PDF, DOCX)
    summary: string;           // Summary of the document's contents
    id: string;                // Additional identifier
    __partitionkey: string;    // Partition key (specific to your data structure)
    // index_key: string
    // metadata_storage_path: string
    // metadata_storage_name: string
    // metadata_storage_size: number
    // metadata_storage_content_md5: string
    // content_size: number
    // content_encoding: any
    // description: any
    // creation_date: string
    // last_modified: string
    // processing_date: any
    // source_processing_date: string
    // source_last_modified: string
    // key_phrases: string[]
    // topics: any[]
    // organizations: string[]
    // persons: string[]
    // locations: string[]
    // cities: string[]
    // countries: string[]
    // language: string
    // translated_language: string
    // paragraphs_count: number
    // summary: string[]
    // categories: any[]
    // captions: any[]
    // title: string
    // translated_title: string
    // author: string
    // content_type: string
    // content_group: string
    page_number: number
    // page_count: any
    // slide_count: any
    // links: any[]
    // emails: any[]
    // //		Index_key -> document Id
    // // 		imageUrl
    // // 		Filename
    // // 		Filelocation
    // // 		Tags
    // // 		Uploadtime
    // // 		LatestProcessTime
    // //    Status
    // // Chat history id per document?
    // document_id: string
    // document_filename: string
    document_url: string
    // document_segments: string[]
    // markets: any[]
    // competitions: any[]
    // technologies: any[]
    // user_keywords: any[]
    // user_categories: any[]
    // user_tags: any[]
    // strategies: any[]
    // tables: string[]
    // tables_count: number
    // kvs: string
    // kvs_count: number
    // geolocation: any
    // restricted: boolean
    // parent_id: any
    // chunk: any
    // vector: any[]
    // parent: Parent
    // image: Image
    // email: any
    // document: Document2
  }
// ...
  export interface Tokens {
    documents: string
    images: string
    translation: string
  }